requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
groq>=0.11.0
httpx>=0.27.0

# FastAPI server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Audio processing
pydub>=0.25.0

# Browser automation and video capture
playwright>=1.40.0
Pillow>=10.0.0

# Async utilities
aiofiles>=23.0.0

# Cloud storage
supabase>=2.0.0
